smart-review-analyzer/requirements.txt
pandas
scikit-learn
streamlit
google-generativeai
python-dotenv
